refactor cognitive agent
